+++ gitpandas/utilities/__init__.py
@@ -1,0 +1,11 @@
+"""
+.. module:: utilities
+   :platform: Unix, Windows
+   :synopsis: Helper methods for plotting or otherwise manipulating output from gitpandas objects
+
+.. moduleauthor:: Will McGinnis <will@pedalwrencher.com>
+
+
+"""
+
+__author__ = 'willmcginnis'